update for semi-working status
angry change of names for machine names.

changed min temp from 170 to 5, this is not desirable for some people, most people will load while hot, but for purposes of testing extruder i set it to 5.

defined driver types for extruders.

changed to 78.7402 based on a config i found on reddit from someone who had calibrated this to be the steps for x and y to be "accurate", haven't tested this yet.

changed inversions of axes.

inverted extruder as it was attempting to load backwards.